apps/agents/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
groq==0.5.0
pydantic==2.13.2
python-dotenv==1.0.0
httpx==0.25.2
supabase==2.3.4
stripe==7.8.0
google-cloud-secret-manager==2.18.2
